+++ src/main/java/damon/backend/repository/community/CommunityLikeRepository.java
@@ -1,7 +1,6 @@
 package damon.backend.repository.community;
 
-import damon.backend.entity.Community;
-import damon.backend.entity.CommunityLike;
+import damon.backend.entity.community.CommunityLike;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -14,13 +13,4 @@ import java.util.Optional;
 @Repository
 public interface CommunityLikeRepository extends JpaRepository<CommunityLike, Long> {
 
-    @EntityGraph(attributePaths = {"community", "member"})
-    @Query("SELECT l FROM CommunityLike l WHERE l.community.communityId = :communityId")
-    List<CommunityLike> findAllFetch(@Param("communityId") Long communityId);
-
-    @EntityGraph(attributePaths = {"community", "member"})
-    @Query("SELECT l FROM CommunityLike l WHERE l.community.communityId = :communityId AND l.member.id = :memberId")
-    Optional<CommunityLike> findOneFetch(@Param("communityId") Long communityId, @Param("memberId") String memberId);
-
-    boolean existsByCommunityCommunityIdAndMemberId(Long communityId, String memberId);
 }
