+++ src/main/java/damon/backend/repository/community/CommunityRepository.java
@@ -1,13 +1,11 @@
 package damon.backend.repository.community;
 
-import damon.backend.entity.Community;
+import damon.backend.entity.community.Community;
 import damon.backend.enums.CommunityType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -16,15 +14,18 @@ import java.util.Optional;
 @Repository
 public interface CommunityRepository extends JpaRepository<Community, Long>  {
 
-    @EntityGraph(attributePaths = {"member", "comments", "likes"})
-    @Query("SELECT c FROM Community c")
-    List<Community> findAllFetch();
+    @Query("SELECT c FROM Community c JOIN FETCH c.member WHERE c.communityId = :communityId")
+    Optional<Community> findOne(Long communityId);
 
-    @EntityGraph(attributePaths = {"member", "comments", "likes"})
-    @Query("SELECT c FROM Community c WHERE c.communityId = :communityId")
-    Optional<Community> findOneFetch(@Param("communityId") Long communityId);
+    @Query("SELECT c FROM Community c JOIN FETCH c.member WHERE c.type = :type ORDER BY c.createdDate DESC")
+    List<Community> findAllByList(CommunityType type);
 
-    @EntityGraph(attributePaths = {"member", "comments", "likes"})
-    @Query("SELECT c FROM Community c WHERE c.type = :communityType")
-    Page<Community> findAllFetchPaging(@Param("communityType") CommunityType communityType, Pageable pageable);
+    @Query("SELECT c FROM Community c JOIN FETCH c.member WHERE c.type = :type ORDER BY c.createdDate DESC")
+    Page<Community> findAllByPage(CommunityType type, Pageable pageable);
+
+    @Query("SELECT c FROM Community c JOIN FETCH c.member WHERE c.type = :type ORDER BY c.createdDate DESC LIMIT 5")
+    List<Community> findTop5ByList(CommunityType type);
+
+    @Query("SELECT c FROM Community c JOIN FETCH c.member m WHERE c.type = :type AND m.id = :memberId ORDER BY c.createdDate DESC")
+    Page<Community> findMyByPage( String memberId, CommunityType type, Pageable pageable);
 }
